add LabelManagementPage
